Replace JHU EP logo with university logo
This change set replaces the JHU EP pseudo-logo with the official
Johns Hopkins University logo. In addition, other references to the
EP program (e.g., in the footer) are replaced with the analogous
references to the university.

Embedding the university logo within the dtx was difficult: the logo
was a Windows-format EPS file that required modification before
embedding. A successful strategy was to find and remove all lines
containing \r\n (^M); these lines appear related to a TIFF or WMF
preview that is itself embedded in the EPS file. In addition, the
first line is modified to be
    %!PS-Adobe-3.1 EPSF-3.0
which adheres to the standard EPS header. (The original header
identified the file as a Windows-format EPS file.)

Due to the difficulty working with the white logo, it would be
advantageous to create an inverted color scheme for slides; the
inverted color scheme would use the blue logo on a white background.