Reduce cropping of logo
According to JHU's identity guidelines (identity.jhu.edu), the
minimum clear space surrounding the logo is the height of the 'H' in
Hopkins. Although this distance is not maintained even with this
change, it does increase the clear space around the logo. The logo
graphic (i.e., EPS image) includes the recommended clear space (the
height of the shield), which is much too large for the available
space in the slides.