Replace "BlueJay" color with PMS 228
This changeset removes the "BlueJay" color in the slides and replaces
it with PMS 288, which is the dark blue used in the Johns Hopkins
logos.

Additional colors from the Johns Hopkins color palette should be
incorporated into a (separate) color theme. See identity.jhu.edu for
additional colors that could be included.