Add JHU theme that reuses color and outer themes
This changeset updates the existing Beamer theme for Johns Hopkins
University to use the bluejay color theme and logo outer theme. The
result is a substantially simpler package that loads the other themes
and specifies the university logo.

This changeset removes the specialized themes for the Whiting School
of Engineering and the Applied Physics Laboratory. Future changes
should create new themes for these divisions.