Remove extraneous text from documentation
This changeset removes instances of "for compatibility with earlier
versions" from the package documentation. This text appeared because
`docstrip' removed comment characters from the Pygment's style
definitions (*.pygstyle) leaving the comment text to be included in
the documentation.

The theme-doc package patches the minted commands (specifically, the
inputminted macro) to write the style definitions to the auxiliary
file, which causes them to be loaded without interference from
`docstrip'. See https://goo.gl/wUd26p for a similar issue that
affected minted's own documentation.