Add examples of Beamer themes
This changeset adds two images of a presentation styled with the theme
to the theme's documentation. It makes it easier to see the theme and
the immediate effect of any changes to the theme.

The example presentation reuses the same example presentation as the
themes shown in Beamer's documentation. beamerthemeexamplebase.tex is
copied (unmodified) from Beamer's official repository
(https://goo.gl/Psoxm8), changeset 3dd33d7.

The testcolor macro is added to the theme-doc package to replace
duplicate definitions in the bluejay and wahoo color themes.