Reorder entries in footline
This changeset alters the content of the foot line so it contains the
(short) date, the (short) title, the (short) author, and the slide
number. The (short) institute is no longer included in the foot line.

To accommodate this modification, the space allocated from the logo
is reduced from 40% of the frame head to 34% of the frame head. The
space is reallocated to the section and subsection navigation.

The logos of various themes are updated to accommodate this smaller
size.