Use key-value options for logo outer theme
This changeset uses pgfopts to parse the package options for the logo
outer theme. Using key-value pairs for the options allows greater
flexibility for future additions. Existing themes that use this
package (e.g., the jhu.edu and seas.virginia.edu themes) are updated
to use the new syntax for the package options.

Backwards compatibility is *not* preserved: all options must be
specified as key-value pairs. Existing uses of this package should be
updated as follows:
- right => position=right
- tree  => base=tree