Place package inside directory for distribution
For distribution, CTAN prescribes the structure of packages. In
particular, all files should be inside a directory that is inside a
ZIP archive. That is, the contents of the distribution archive should
be as follows:

  my-package.zip:
    my-package/
      README
      my-package.dtx
      my-package.ins
      my-package.pdf

This change follows that recommendation instead of having all the
files in the root of the distribution archive (i.e., no internal
directory structure).

CTAN instructions: https://ctan.org/help/upload-pkg