Add prerequisite on package to documentation
This change adds a missing prerequisite on the package (e.g., *.sty
file) to the documentation for that package.

Previously, the Makefile for each package listed the package and its
documentation in the default targets, which masked this issue because
the package was always built prior to the documentation, but building
the documentation by itself could fail due to the missing
prerequisite.