Correct search path for identically-named packages
This change clears the search path (i.e., the `vpath` directive) for
style files for the packages related to syllabi. Altering the search
path is necessary because the existence of identically-named packages
causes make to believe that the prerequisite is already satisfied.

This fix is not an ideal long-term solution: All packages in this
repository should have unique names, which would eliminate the need to
tweak the search path.