Use consistent rule for building packages (#12)
This change removes the order-only prerequisites used to build other
packages (in this repository) in favor of "normal" prerequisites.

For identically-named packages (e.g., the "syllabus" packages for the
Whiting School of Engineering and Engineering for Professionals (EP) 
program), the search path (i.e., the `vpath` directive) is cleared so
that make doesn't erroneously believe that the prerequisite is already
satisfied. This workaround is not an ideal long-term solution: all 
packages in this repository should have unique names.