Use Clear Sans as the default typeface
The Microsoft PowerPoint template for Engineering for Professionals
(EP) uses Tahoma as the typeface. Unfortunately, there does not appear
to be any metrically compatible typeface. Verdana was derived from
Tahoma and is metrically compatible with Vera Sans (Arev and Bera
Sans) and DejaVu Sans, but Clear Sans appears to be a closer match for
Tahoma due to its different forms for uppercase 'I' and lowercase 'l'
and slightly narrower width. Hence, this change uses Clear Sans as the
typeface for the ep.jhu.edu presentation theme.

References #17