Use Helvetica for examples
Removing the times package results in the use of Computer Modern,
which not ideal for presentations. In addition, using Computer Modern
causes a line break in "Engineering for Professionals" in the
ep.jhu.edu presentation theme, which is not appropriate. Hence, this
change specifies the use of Helvetica for all examples.