Remove times package from example (#16)
The inclusion of the times package prevents examples of Beamer
presentation themes from using a typeface specified by the theme.
Hence, this change removes the use of the times package (see also
josephwright/beamer#629).

Removing the times package results in the use of Computer Modern,
which not ideal for presentations. In addition, using Computer Modern
causes a line break in "Engineering for Professionals" in the
ep.jhu.edu presentation theme, which is not appropriate. Hence, this
change specifies the use of Helvetica for all examples.