Patch epstopdf to search for EPS images
Because EPS images are converted to PDFs rather than being embedded
directly in documents, `epstopdf` may not be able to find the images,
such as when the images are found using the TEXINPUTS environment
variable (e.g., when TEXINPUTS references a local texmf directory).
This change patches epstopdf to use `kpsewhich` to find EPS images,
which eliminates the need to place images used by the ep.jhu.edu
Beamer presentation theme (e.g., logos) in the same directory as the
LaTeX document.