Only mark external files as secondary targets
External files (i.e., those fetched on-demand) are not deleted
automatically because they are expected to be used again. This change
guards the declaration of the .SECONDARY target so any external files
are its prerequisites. Otherwise, all targets are treated as secondary
targets, and no target is removed even when it is an intermediate.

(cherry picked from commit joel-coffman/latex-incubator@308dacb)