Update EP Beamer theme to match official template (#20)
Engineering for Professionals (EP) released an updated Microsoft
PowerPoint template designed for online courses. This change updates
the ep.jhu.edu Beamer presentation theme so it is consistent with that
template.

The Microsoft PowerPoint template uses Tahoma as the typeface.
Unfortunately, there does not appear to be any metrically compatible
typeface. Verdana was derived from Tahoma and is metrically compatible
with Vera Sans (Arev and Bera Sans) and DejaVu Sans, but Clear Sans is
a closer match for Tahoma due to its different forms for uppercase 'I'
and lowercase 'l' and slightly narrower width. Hence, this change uses
Clear Sans as the default typeface for the theme.

Because EPS images are converted to PDFs rather than being embedded
directly in documents, `epstopdf` may not be able to find the images,
such as when the images are found using the TEXINPUTS environment
variable (e.g., when TEXINPUTS references a local texmf directory).
This change patches epstopdf to use `kpsewhich` to find EPS images,
which eliminates the need to place images used by the ep.jhu.edu
Beamer presentation theme (e.g., logos) in the same directory as the
LaTeX document.

Closes #17